requirements.txt
flask
requests
gunicorn
sqlalchemy
python-dotenv
reportlab
